requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
google-cloud-storage
sse-starlette
